Update login.py - Conversion to Class
The original boiler plate login.py did not use a class system; rewriting from scratch. Hosting all login-functions within a class allows for the login protocol to be called: from login import Login, which will execute the login methods appropriate for the user & simplify the code for the app itself.

1 - added a class initialization for the attributes needed: a list of existing user codenames will be required (read_users) function not written yet. Each user will have a codename, password, & last name; no assigned value.

2 - prompt_user function -> there are 2 sign in methods & 3 outcomes. The user has a codename. The user does not have a codename but wants to create one. the user does not have a codename but wants to not become an agent (exits program).

3 - code block written for new_user login method -> user inputs multiple values with are transformed to a string of uppercase characters & numbers with no spaces. Same for their password. If the codename created exists, throw error.
If the codename doesn't exist, we collect a last name. This provides assigned values to all of the original class attributes.

4 - code comments note which functions have not been written yet / code can't function without them.